chore: platform destroyer; - improve output
